requirements.txt
discord.py==2.4.0
pytz==2023.3
PyNaCl==1.5.0
aiohttp~=3.10.1